Update
BugFix on notifications.
Changed 'stop test' buttons, so that the user has to push twice to stop
a test.
Changed end of word recog end. Now a canvas shows up when the user has
chosen the needed amount of words and asks if the user wants to change
something with 5 seconds of time or just stop.